[mem & norm] adapting to norm when necessary memory cleaned
