+++ Makefile
@@ -59,6 +59,7 @@ MANDATORY_FILES =	${CHECKS_PATH}/00_errors.c			\
 					${WIND}/image.c						\
 					${WIND}/draw.c						\
 					mainfer.c							\
+					exit.c								\
 					# main.c							\
 
 #headers
